requirements-test.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
channels==4.0.0
charset-normalizer==3.2.0
colorama==0.4.6
contourpy==1.1.0
coverage==7.2.1
cycler==0.11.0
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.1.1
django-allauth==0.54.0
django-cleanup==8.0.0
django-colorfield==0.9.0
django-cors-headers==4.2.0
django-environ==0.9.0
django-filter==23.2
djangochannelsrestframework==1.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.7
easyocr==1.7.0
editdistance==0.6.2
efficientnet==1.0.0
essential-generators==1.0
filelock==3.12.2
fonttools==4.42.0
fsspec==2023.6.0
gensim==4.3.1
h5py==3.9.0
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.1
imgaug==0.4.0
importlib-resources==6.0.1
inflection==0.5.1
Jinja2==3.1.2
Keras-Applications==1.0.8
keras-ocr==0.9.2
kiwisolver==1.4.4
lao-test==1.1.0
LaoNLP==0.7
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
ninja==1.11.1
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.7
pyclipper==1.3.0.post4
PyJWT==2.8.0
pyparsing==3.0.9
pytesseract==0.3.10
pythainlp==4.0.2
python-bidi==0.4.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
scikit-image==0.21.0
scipy==1.10.1
shapely==2.0.1
six==1.16.0
smart-open==6.3.0
sorl-thumbnail==12.9.0
sorl-thumbnail-serializer-field==0.2.1
sqlparse==0.4.4
sympy==1.12
tifffile==2023.7.10
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
validators==0.21.2
zipp==3.16.2
